+++ resources/js/Layouts/ChatView.jsx
@@ -1,0 +1,7 @@
+export default function ChatView() {
+    return (
+        <div className="h-96">
+
+        </div>
+    )
+}
